Move the filtering strategies into directory/module namespaces
