+++ lib/active_set/filtering/operation.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require_relative '../attribute_instruction'
-require_relative './enumerable_strategy'
-require_relative './active_record_strategy'
+require_relative './enumerable/strategy'
+require_relative './active_record/strategy'
 
 class ActiveSet
   module Filtering
@@ -17,7 +17,7 @@ class ActiveSet
                                  .map { |k, v| AttributeInstruction.new(k, v) }
 
         activerecord_filtered_set = attribute_instructions.reduce(@set) do |set, attribute_instruction|
-          maybe_set_or_false = ActiveRecordStrategy.new(set, attribute_instruction).execute
+          maybe_set_or_false = ActiveRecord::Strategy.new(set, attribute_instruction).execute
           next set unless maybe_set_or_false
 
           attribute_instruction.processed = true
@@ -27,7 +27,7 @@ class ActiveSet
         return activerecord_filtered_set if attribute_instructions.all?(&:processed?)
 
         attribute_instructions.reject(&:processed?).reduce(activerecord_filtered_set) do |set, attribute_instruction|
-          maybe_set_or_false = EnumerableStrategy.new(set, attribute_instruction).execute
+          maybe_set_or_false = Enumerable::Strategy.new(set, attribute_instruction).execute
           next set unless maybe_set_or_false
 
           attribute_instruction.processed = true
